Add Semester & modified edu_sys post functions

+++ SunReadSource/src/main/java/com/syzton/sunread/service/education_system/EduGroupRepositoryService.java
@@ -1,12 +1,10 @@
 package com.syzton.sunread.service.education_system;
 
 import com.syzton.sunread.dto.education_system.EduGroupDTO;
-import com.syzton.sunread.exception.coinhistory.CoinHistoryNotFoundException;
-import com.syzton.sunread.model.education_system.Clazz;
 import com.syzton.sunread.model.education_system.EduGroup;
 import com.syzton.sunread.repository.education_system.EduGroupRepository;
-
-import javassist.NotFoundException;
+import com.syzton.sunread.exception.common.DuplicateException;
+import com.syzton.sunread.exception.common.NotFoundException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,7 +31,13 @@ public class EduGroupRepositoryService implements EduGroupService {
     @Override
     public EduGroup add(EduGroupDTO add) {
 
-        LOGGER.debug("Adding a new clazz entry with information: {}", add);
+        LOGGER.debug("Adding a new edu group entry with information: {}", add);
+        
+        EduGroup exits = repository.findOne(add.getId());        
+        if(exits != null){
+            throw new DuplicateException("EduGroup with name: "+add.getName()+" is already exits..");
+        }
+        
         EduGroup model = EduGroup.getBuilder(add.getName())
         		.description(add.getDescription()).build();   
         return repository.save(model);
@@ -42,49 +46,53 @@ public class EduGroupRepositoryService implements EduGroupService {
 
     @Transactional(rollbackFor = {NotFoundException.class})
     @Override
-    public EduGroup deleteById(Long id) throws NotFoundException {
-        LOGGER.debug("Deleting a clazz entry with id: {}", id);
+    public EduGroup deleteById(Long id){
+        LOGGER.debug("Deleting a edu group entry with id: {}", id);
 
         EduGroup deleted = findById(id);
-        LOGGER.debug("Deleting clazz entry: {}", deleted);
-
+        if(deleted == null)
+            throw new NotFoundException("No Edu Group found with id: " + id);
+        
+        LOGGER.debug("Deleting edu group entry: {}", deleted);
         repository.delete(deleted);
         return deleted;
     }
 
     @Transactional(rollbackFor = {NotFoundException.class})
     @Override
-    public EduGroup update(EduGroupDTO updated)throws  NotFoundException{
+    public EduGroup update(EduGroupDTO updated){
         LOGGER.debug("Updating contact with information: {}", updated);
 
         EduGroup model = findById(updated.getId());
+        if(model == null)
+            throw new NotFoundException("No Edu Group found with id: " + updated.getId());
         LOGGER.debug("Found a note entry: {}", model);
 
         model.update(updated.getName());
         return model;
     }
 
-    @Transactional(readOnly = true, rollbackFor = {NotFoundException.class})
     @Override
-    public EduGroup findById(Long id) throws NotFoundException {
-        LOGGER.debug("Finding a clazz entry with id: {}", id);
+    public EduGroup findById(Long id) {
+        LOGGER.debug("Finding a edu group entry with id: {}", id);
 
         EduGroup found = repository.findOne(id);
-        LOGGER.debug("Found clazz entry: {}", found);
+        LOGGER.debug("Found edu group entry: {}", found);
 
         if (found == null) {
-            throw new NotFoundException("No clazz entry found with id: " + id);
+            throw new NotFoundException("No edu group entry found with id: " + id);
         }
 
         return found;
     }
 
-
-
-    @Transactional(rollbackFor = {NotFoundException.class})
     @Override
-    public Page<EduGroup> findAll(Pageable pageable) throws NotFoundException {
-        LOGGER.debug("Finding all clazz entries");
-        return repository.findAll(pageable);
+    public Page<EduGroup> findAll(Pageable pageable){
+        LOGGER.debug("Finding all edu Group entries");
+        Page<EduGroup> eduPages = repository.findAll(pageable);
+        if (eduPages == null) {
+            throw new NotFoundException("No edu group entry found");
+        }
+        return eduPages;
     }
 }
